refactor: 공통으로 사용하는 response dto 패키지 위치 변경
